+++ CotEditor/Sources/EditorTextView+Commenting.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2018 1024jp
+//  © 2014-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -47,7 +47,7 @@ extension EditorTextView: Commenting {
     /// toggle comment state in selection
     @IBAction func toggleComment(_ sender: Any?) {
         
-        if self.canUncomment(range: self.selectedRange, partly: false) {
+        if self.canUncomment(partly: false) {
             self.uncomment(types: .both, fromLineHead: self.commentsAtLineHead)
         } else {
             self.commentOut(types: .both, fromLineHead: self.commentsAtLineHead)
@@ -118,27 +118,35 @@ extension Commenting where Self: NSTextView {
         
         guard self.blockCommentDelimiters != nil || self.inlineCommentDelimiter != nil else { return }
         
-        let string = self.string
-        
-        guard
-            let selectedRange = Range(self.selectedRange, in: string),
-            let targetRange = self.commentingRange(fromLineHead: fromLineHead)
-            else { return }
-        
         let spacer = self.appendsCommentSpacer ? " " : ""
-        var new: (String, NSRange)?
+        let targetRanges = self.commentingRanges(fromLineHead: self.commentsAtLineHead)
+        let infos: [(string: String, location: Int, before: Bool)] = {
+            if let delimiter = self.inlineCommentDelimiter, types.contains(.inline) {
+                return self.string.inlineCommentOut(delimiter: delimiter, spacer: spacer, ranges: targetRanges)
+            }
+            if let delimiters = self.blockCommentDelimiters, types.contains(.block) {
+                return self.string.blockCommentOut(delimiters: delimiters, spacer: spacer, ranges: targetRanges)
+            }
+            return []
+        }()
         
-        // insert delimiters
-        if let delimiter = self.inlineCommentDelimiter, types.contains(.inline) {
-            new = string.inlineCommentOut(delimiter: delimiter, spacer: spacer, range: targetRange, selectedRange: selectedRange)
-            
-        } else if let delimiters = self.blockCommentDelimiters, types.contains(.block) {
-            new = string.blockCommentOut(delimiters: delimiters, spacer: spacer, range: targetRange, selectedRange: selectedRange)
-        }
+        guard !infos.isEmpty else { return }
         
-        guard let (newString, newSelectedRange) = new else { return }
+        let newStrings = infos.map { $0.string }
+        let replacementRanges = infos.map { NSRange($0.location..<$0.location) }
+        let selectedRanges = (self.rangesForUserTextChange ?? self.selectedRanges)
+            .map { $0.rangeValue }
+            .map { range -> NSRange in
+                let location = infos
+                    .filter { $0.location < range.location || ($0.location == range.location && $0.before) }
+                    .reduce(into: range.location) { $0 += ($1.string as NSString).length }
+                let length = infos.filter { range.lowerBound < $0.location && $0.location < range.upperBound }
+                    .reduce(into: range.length) { $0 += ($1.string as NSString).length }
+                
+                return NSRange(location: location, length: length)
+            }
         
-        self.replace(with: newString, range: NSRange(targetRange, in: string), selectedRange: newSelectedRange,
+        self.replace(with: newStrings, ranges: replacementRanges, selectedRanges: selectedRanges,
                      actionName: "Comment Out".localized)
     }
     
@@ -148,55 +156,64 @@ extension Commenting where Self: NSTextView {
         
         guard self.blockCommentDelimiters != nil || self.inlineCommentDelimiter != nil else { return }
         
-        let string = self.string
-        
-        guard
-            let selectedRange = Range(self.selectedRange, in: string),
-            let targetRange = self.commentingRange(fromLineHead: fromLineHead),
-            !targetRange.isEmpty else { return }
-        
         let spacer = self.appendsCommentSpacer ? " " : ""
-        var new: (String, NSRange)?
+        let targetRanges = self.commentingRanges(fromLineHead: self.commentsAtLineHead).filter { $0.length > 0 }
+        let indices: IndexSet = {
+            if let delimiters = self.blockCommentDelimiters, types.contains(.block) {
+                let indices = self.string.blockUncomment(delimiters: delimiters, spacer: spacer, ranges: targetRanges)
+                if !indices.isEmpty { return indices }
+            }
+            if let delimiter = self.inlineCommentDelimiter, types.contains(.inline) {
+                return self.string.inlineUncomment(delimiter: delimiter, spacer: spacer, ranges: targetRanges)
+            }
+            return IndexSet()
+        }()
         
-        if let delimiters = self.blockCommentDelimiters, types.contains(.block) {
-            new = string.blockUncomment(delimiters: delimiters, spacer: spacer, range: targetRange, selectedRange: selectedRange)
-        }
-        if let delimiter = self.inlineCommentDelimiter, types.contains(.inline), new == nil {
-            new = string.inlineUncomment(delimiter: delimiter, spacer: spacer, range: targetRange, selectedRange: selectedRange)
-        }
+        guard !indices.isEmpty else { return }
         
-        guard let (newString, newSelectedRange) = new else { return }
+        let replacementRanges = indices.rangeView.map { NSRange($0.lowerBound..<$0.upperBound) }
+        let newStrings = [String](repeating: "", count: replacementRanges.count)
+        let selectedRanges = (self.rangesForUserTextChange ?? self.selectedRanges)
+            .map { $0.rangeValue }
+            .map { range -> NSRange in
+                let location = range.location - indices.count { $0 < range.location }
+                let length = range.length - indices.count { range.contains($0) }
+                
+                return NSRange(location: location, length: length)
+            }
         
-        self.replace(with: newString, range: NSRange(targetRange, in: string), selectedRange: newSelectedRange,
+        
+        self.replace(with: newStrings, ranges: replacementRanges, selectedRanges: selectedRanges,
                      actionName: "Uncomment".localized)
     }
     
     
     /// whether given range can be uncommented
-    func canUncomment(range: NSRange, partly: Bool) -> Bool {
+    func canUncomment(partly: Bool) -> Bool {
         
         guard self.blockCommentDelimiters != nil || self.inlineCommentDelimiter != nil else { return false }
         
-        guard
-            let targetRange = self.commentingRange(fromLineHead: self.commentsAtLineHead),
-            !targetRange.isEmpty else { return false }
+        let targets = self.commentingRanges(fromLineHead: self.commentsAtLineHead)
+            .filter { $0.length > 0 }
+            .map { (self.string as NSString).substring(with: $0) }
         
-        let target = self.string[targetRange]
+        guard !targets.isEmpty else { return false }
         
         if let delimiters = self.blockCommentDelimiters {
-            if target.hasPrefix(delimiters.begin), target.hasSuffix(delimiters.end) {
+            let predicate: ((String) -> Bool) = { $0.hasPrefix(delimiters.begin) && $0.hasSuffix(delimiters.end) }
+            
+            if partly ? targets.contains(where: predicate) : targets.allSatisfy(predicate) {
                 return true
             }
         }
         
         if let delimiter = self.inlineCommentDelimiter {
-            let lines = target.components(separatedBy: "\n")
-            var commentLineCount = 0
-            for line in lines where line.hasPrefix(delimiter) {
-                if partly { return true }
-                commentLineCount += 1
+            let predicate: ((String) -> Bool) = { $0.hasPrefix(delimiter) }
+            let lines = targets.flatMap { $0.components(separatedBy: "\n") }
+            
+            if partly ? lines.contains(where: predicate) : lines.allSatisfy(predicate) {
+                return true
             }
-            return commentLineCount == lines.count
         }
         
         return false
@@ -207,117 +224,67 @@ extension Commenting where Self: NSTextView {
     // MARK: Private Methods
     
     /// return commenting target range
-    private func commentingRange(fromLineHead: Bool) -> Range<String.Index>? {
+    private func commentingRanges(fromLineHead: Bool) -> [NSRange] {
         
-        guard let selectedRange = Range(self.selectedRange, in: self.string) else { return nil }
-        
-        return fromLineHead ? self.string.lineRange(for: selectedRange, excludingLastLineEnding: true) : selectedRange
+        return (self.rangesForUserTextChange ?? [])
+            .map { $0.rangeValue }
+            .map { fromLineHead ? self.string.lineRange(for: $0, excludingLastLineEnding: true) : $0 }
+            .unique
     }
     
 }
 
 
 
-private extension StringProtocol where Self.Index == String.Index {
+private extension String {
     
     /// append inline style comment delimiters in range inserting spacer after delimiters and return commented-out string and new selected range
-    func inlineCommentOut(delimiter: String, spacer: String, range: Range<Index>, selectedRange: Range<Index>) -> (String, NSRange) {
+    func inlineCommentOut(delimiter: String, spacer: String, ranges: [NSRange]) -> [(string: String, location: Int, before: Bool)] {
         
-        let target = self[range]
+        let regex = try! NSRegularExpression(pattern: "^", options: [.anchorsMatchLines])
         
-        let newString = delimiter + spacer + target.replacingOccurrences(of: "\n", with: "\n" + delimiter + spacer)
-        let cursorOffset = newString.count - target.count
-        
-        let newSelectedRange = self.seletedRange(range: range, selectedRange: selectedRange, newString: newString, offset: cursorOffset)
-        
-        return (newString, newSelectedRange)
+        return ranges.flatMap { regex.matches(in: self, range: $0) }
+            .map { $0.range.location }
+            .unique
+            .map { (delimiter + spacer, $0, true) }
     }
     
     
     /// append block style comment delimiters in range inserting spacer between string and delimiters and return commented-out string and new selected range
-    func blockCommentOut(delimiters: Pair<String>, spacer: String, range: Range<Index>, selectedRange: Range<Index>) -> (String, NSRange) {
+    func blockCommentOut(delimiters: Pair<String>, spacer: String, ranges: [NSRange]) -> [(string: String, location: Int, before: Bool)] {
         
-        let target = self[range]
-        
-        let newString = delimiters.begin + spacer + target + spacer + delimiters.end
-        let cursorOffset = delimiters.begin.count + spacer.count
-        
-        let newSelectedRange = self.seletedRange(range: range, selectedRange: selectedRange, newString: newString, offset: cursorOffset)
-        
-        return (newString, newSelectedRange)
+        return ranges.flatMap { [(delimiters.begin + spacer, $0.lowerBound, true), (spacer + delimiters.end, $0.upperBound, false)] }
     }
     
     
-    /// remove inline style comment delimiters in range removing also spacers after delimiter and return uncommented string and new selected range
-    func inlineUncomment(delimiter: String, spacer: String, range: Range<Index>, selectedRange: Range<Index>) -> (String, NSRange)? {
+    /// find inline style comment delimiters with and without spacers in range and return a character location IndexSet to remove
+    func inlineUncomment(delimiter: String, spacer: String, ranges: [NSRange]) -> IndexSet {
         
-        let target = self[range]
+        let pattern = NSRegularExpression.escapedPattern(for: delimiter)
+        let spacerPattern = spacer.isEmpty ? "" : "(?:" + spacer + ")?"
+        let regex = try! NSRegularExpression(pattern: "^" + pattern + spacerPattern,
+                                             options: [.anchorsMatchLines])
         
-        let lines = target.components(separatedBy: "\n")
-        let newLines: [String] = lines.map { line in
-            guard line.hasPrefix(delimiter) else { return line }
-            
-            let newLine = line.dropFirst(delimiter.count)
-            
-            guard !spacer.isEmpty, newLine.hasPrefix(spacer) else { return String(newLine) }
-            
-            return String(newLine.dropFirst(spacer.count))
-        }
-        
-        let newString = newLines.joined(separator: "\n")
-        let cursorOffset = -(target.count - newString.count)
-        
-        guard cursorOffset != 0 else { return nil }
-        
-        let newSelectedRange = self.seletedRange(range: range, selectedRange: selectedRange, newString: newString, offset: cursorOffset)
-        
-        return (newString, newSelectedRange)
+        return ranges.flatMap { regex.matches(in: self, range: $0) }
+            .map { $0.range }
+            .map { $0.lowerBound..<$0.upperBound }
+            .reduce(into: IndexSet()) { $0.insert(integersIn: $1) }
     }
     
     
     /// remove block style comment delimiters in range removing also spacers between string and delimiter and return uncommented string and new selected range
-    func blockUncomment(delimiters: Pair<String>, spacer: String, range: Range<Index>, selectedRange: Range<Index>) -> (String, NSRange)? {
+    func blockUncomment(delimiters: Pair<String>, spacer: String, ranges: [NSRange]) -> IndexSet {
         
-        let target = self[range]
+        let beginPattern = NSRegularExpression.escapedPattern(for: delimiters.begin)
+        let endPattern = NSRegularExpression.escapedPattern(for: delimiters.end)
+        let spacerPattern = spacer.isEmpty ? "" : "(?:" + spacer + ")?"
+        let regex = try! NSRegularExpression(pattern: "\\A(" + beginPattern + spacerPattern + ").*?(" + spacerPattern + endPattern + ")\\Z",
+                                             options: [.dotMatchesLineSeparators])
         
-        guard target.hasPrefix(delimiters.begin), target.hasSuffix(delimiters.end) else { return nil }
-        
-        let trimFrom = target.index(target.startIndex, offsetBy: delimiters.begin.count)
-        let trimTo = target.index(target.endIndex, offsetBy: -delimiters.end.count)
-        var substring = target[trimFrom..<trimTo]
-        var cursorOffset = -delimiters.begin.count
-        
-        if !spacer.isEmpty, substring.hasPrefix(spacer), substring.hasSuffix(spacer) {
-            let trimFrom = substring.index(substring.startIndex, offsetBy: spacer.count)
-            let trimTo = substring.index(substring.endIndex, offsetBy: -spacer.count)
-            substring = substring[trimFrom..<trimTo]
-            cursorOffset -= spacer.count
-        }
-        
-        let newString = String(substring)
-        let newSelectedRange = self.seletedRange(range: range, selectedRange: selectedRange, newString: newString, offset: cursorOffset)
-        
-        return (newString, newSelectedRange)
-    }
-    
-    
-    
-    // MARK: Private Methods
-    
-    /// create selected range after commenting-out/uncommenting
-    private func seletedRange(range: Range<Index>, selectedRange: Range<Index>, newString: String, offset: Int) -> NSRange {
-        
-        var nsRange: NSRange
-        
-        if selectedRange.isEmpty {
-            nsRange = NSRange(selectedRange, in: self)
-            nsRange.location += offset
-        } else {
-            nsRange = NSRange(range, in: self)
-            nsRange.length = newString.utf16.count
-        }
-        
-        return nsRange
+        return ranges.flatMap { regex.matches(in: self, range: $0) }
+            .flatMap { [$0.range(at: 1), $0.range(at: 2)] }
+            .map { $0.lowerBound..<$0.upperBound }
+            .reduce(into: IndexSet()) { $0.insert(integersIn: $1) }
     }
     
 }