+++ CotEditor/Sources/NSTextView+TextReplacement.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2018 1024jp
+//  © 2014-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -64,12 +64,11 @@ extension NSTextView {
             self.undoManager?.setActionName(actionName)
         }
         
-        let attributes = self.typingAttributes
         
         textStorage.beginEditing()
-        // use backwards enumeration to skip adjustment of applying location
+        // use a backward enumeration to skip adjustment of applying location
         for (string, range) in zip(strings, ranges).reversed() {
-            let attrString = NSAttributedString(string: string, attributes: attributes)
+            let attrString = NSAttributedString(string: string, attributes: self.typingAttributes)
             
             textStorage.replaceCharacters(in: range, with: attrString)
         }
@@ -88,7 +87,16 @@ extension NSTextView {
     /// set undoable selection change
     func setSelectedRangesWithUndo(_ ranges: [NSValue]) {
         
-        self.selectedRanges = ranges
+        if let self = self as? NSTextView & MultiCursorEditing,
+            let ranges = ranges as? [NSRange],
+            let set = self.prepareForSelectionUpdate(ranges)
+        {
+            self.selectedRanges = set.selectedRanges
+            self.insertionLocations = set.insertionLocations
+            
+        } else {
+            self.selectedRanges = ranges
+        }
         
         self.undoManager?.registerUndo(withTarget: self) { target in
             target.setSelectedRangesWithUndo(ranges)
@@ -103,31 +111,58 @@ extension NSTextView {
     }
     
     
-    /// trim all trailing whitespace with/without keeeping editing point
+    /// trim all trailing whitespace with/without keeping editing point
     func trimTrailingWhitespace(ignoresEmptyLines: Bool, keepingEditingPoint: Bool = false) {
         
         assert(Thread.isMainThread)
 
         let ranges = self.string.rangesOfTrailingWhitespace(ignoresEmptyLines: ignoresEmptyLines)
+        let editingRanges = (self.rangesForUserTextChange ?? self.selectedRanges).map { $0.rangeValue }
         
-        // exclude editing line if needed
-        let replacementRanges: [NSRange] = {
-            guard keepingEditingPoint else { return ranges }
-            
-            let cursorLocation = self.selectedRange.location
-            return ranges.filter { $0.upperBound != cursorLocation && !$0.contains(cursorLocation) }
-        }()
-        
+        // exclude editing lines if needed
+        let replacementRanges: [NSRange] = keepingEditingPoint
+            ? ranges.filter { range in editingRanges.allSatisfy { !$0.touches(range) } }
+            : ranges
+           
         guard !replacementRanges.isEmpty else { return }
         
         let replacementStrings = [String](repeating: "", count: replacementRanges.count)
         
-        self.replace(with: replacementStrings, ranges: replacementRanges, selectedRanges: nil,
+        // calculate selectedRanges after deletion
+        let removedIndexes = replacementRanges.reduce(into: IndexSet()) { $0.insert(integersIn: $1.lowerBound..<$1.upperBound) }
+        let selectedRanges: [NSRange] = editingRanges.map { range in
+            let location = range.location - removedIndexes.count { $0 < range.location }
+            let length = range.length - removedIndexes.count { range.contains($0) }
+            
+            return NSRange(location: location, length: length)
+        }
+        
+        self.replace(with: replacementStrings, ranges: replacementRanges, selectedRanges: selectedRanges,
                      actionName: "Trim Trailing Whitespace".localized)
     }
     
 }
 
+
+
+extension String {
+    
+    /// Create a String from Any but `anyString` must be either String or NSAttributedString.
+    ///
+    /// - Parameter anyString: String or NSAttributedString.
+    init(anyString: Any) {
+        
+        switch anyString {
+        case let string as String:
+            self = string
+        case let attributedString as NSAttributedString:
+            self = attributedString.string
+        default:
+            preconditionFailure()
+        }
+    }
+    
+}
 
 
 extension String {